+++ src/deepymod/data/__init__.py
@@ -1,1 +1,1 @@
-from .base import Dataset, Dataset_2D, DeePyModGPULoader
+from .base import Dataset, Dataset_2D, GPULoader, get_train_test_loader
