+++ src/deepymod/data/samples.py
@@ -2,8 +2,14 @@ import torch
 import numpy as np
 from numpy import ndarray
 from numpy.random import default_rng
-from abc import ABC, ABCMeta
+from abc import ABC, ABCMeta, abstractmethod
 from deepymod.data.base import Subsampler, Dataset
+
+
+class Subsampler(ABC, metaclass=ABCMeta):
+    @abstractmethod
+    def sample():
+        raise NotImplementedError
 
 
 class Subsample_grid(Subsampler):
@@ -45,23 +51,3 @@ class Subsample_random(Subsampler):
         subsampled_coords = torch.tensor(grid[:, :, x_idx].reshape(-1, 2))
         subsampled_data = torch.tensor(grid_data[:, :, x_idx].reshape(-1, 1))
         return subsampled_coords, subsampled_data
-
-
-class MatlabDataset2D(Dataset):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    def load(self):
-        """Output: Grid[N x M x L],  Data[N x M x O],
-        N = Coordinate dimension 0
-        M = Coordinate dimension 1
-        L = Input data dimension
-        O = Output data dimension
-        """
-        x0 = np.linspace(0, 2 * np.pi, 100)
-        x1 = np.linspace(-np.pi, np.pi, 100)
-        X0, X1 = np.meshgrid(x0, x1)
-        y = np.sinc(X0 * X1)
-        coords = torch.tensor(np.stack((X0, X1)))  # .reshape(-1, 2))
-        data = torch.tensor(y).unsqueeze(0)  # .reshape(-1, 1))
-        return coords, data
